feat: weapon fireRate and reload added

+++ dk.sdu.se4.common/src/main/java/dk/sdu/se4/common/entity/part/WeaponPart.java
@@ -16,9 +16,9 @@ public class WeaponPart implements EntityPart {
     private int totalBullets;
     private int magazineSize;
     private int currentMagazine;
-    private int fireRate;
+    private float fireRate;
 
-    public WeaponPart(boolean type, int totalBullets, int magazineSize, int fireRate) {
+    public WeaponPart(boolean type, int totalBullets, int magazineSize, float fireRate) {
         this.type = type;
         this.totalBullets = totalBullets;
         this.magazineSize = magazineSize;
@@ -64,7 +64,7 @@ public class WeaponPart implements EntityPart {
         return currentMagazine + "/" + (totalBullets - currentMagazine);
     }
 
-    public int getFireRate() {
+    public float getFireRate() {
         return fireRate;
     }
